Fix obscure with-eval-in/template related bug
For some reason strings and symbols weren't being properly spliced into
the with-eval-in template expression with ~(str x).  Namespaces were
coerced properly to strings, but if a user supplied a symbol or string
representing a namespace themselves (as via with-logs in tools.logging)
then it broke.

The workaround is to define the functions in the pod and then call them
with .invoke.  I don't know why exactly this works, beyond the fact that
it doesn't involve print/reading anymore the way with-eval-in and
boot.from.backtick/template do.